+++ core/moengage_flutter/test/data_provider/data_provider.dart
@@ -1,0 +1,19 @@
+const num number = 10.56;
+
+const setUserAttributesData = {
+  'bool': true,
+  'int': 10,
+  'double': 10.5,
+  'num': number,
+  'string': 'data',
+  'int-array': [1, 2, 3, 4],
+  'double-array': [1.0, 1.5, 2.456],
+  'num-array': [1.04, 1, 2.456],
+  'string-array': ['data', 'array']
+};
+
+const unsetUserAttributesData = {
+  'obj': {},
+  'bool-array': [true, false],
+  'mixed-array': [1, 1.5, 'data']
+};