+++ core/moengage_flutter/CHANGELOG.md
@@ -3,6 +3,7 @@
 # Next Release Date
 
 ## Next Version
+- Added support for array in user attributes
 - Android 
   - Google Policy - Delete User details API
   - Native SDK Updated to support version `12.10.01` and above