+++ core/moengage_flutter_platform_interface/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MoEngage Flutter Platform Interface
 
+# Next Release Date
+
+## Next Version
+- Added support for array in user attributes
+
 # 13-09-2023
 
 ## 1.0.0