+++ core/moengage_flutter_web/CHANGELOG.md
@@ -1,5 +1,8 @@
 # MoEngage Flutter Web Plugin
+# Next Release
 
+## Next Version
+- Added support for array in user attributes
 # 13-09-2023
 
 ## 2.0.0